RUMM-149 Extract generic abstract FilePeristenceStrategyTest class
